Delete redundant methods in quorum_lachowski.py
